task 1 modification TL

+++ GitProj01/src/main/java/com/nt/service/Arithmetic.java
@@ -2,7 +2,7 @@ package com.nt.service;
 
 public class Arithmetic 
 {
-	public int sum(int x,int y)
+	public double sum(double x,double y)
 	{
 		return x+y;
 	}
